small clean up
